Fix unsigned to u8, needed for windows

+++ part16/include/ppu.h
@@ -45,12 +45,12 @@ typedef struct {
     u8 x;
     u8 tile;
     
-    unsigned f_cgb_pn : 3;
-    unsigned f_cgb_vram_bank : 1;
-    unsigned f_pn : 1;
-    unsigned f_x_flip : 1;
-    unsigned f_y_flip : 1;
-    unsigned f_bgp : 1;
+    u8 f_cgb_pn : 3;
+    u8 f_cgb_vram_bank : 1;
+    u8 f_pn : 1;
+    u8 f_x_flip : 1;
+    u8 f_y_flip : 1;
+    u8 f_bgp : 1;
 
 } oam_entry;
